requirements_cdk.txt
aws-cdk-lib==2.106.1
constructs==10.3.0

